requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.1
bs4==0.0.1
camelot-py==0.8.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.2
Elixir==0.7.1
et-xmlfile==1.0.1
facebook-scraper==0.1.12
fake-useragent==0.1.11
googleDriveFileDownloader==0.0.3
idna==2.9
jdcal==1.4.1
Jinja2==2.11.2
kiwisolver==1.2.0
lxml==4.5.1
MarkupSafe==1.1.1
matplotlib==3.2.1
numpy==1.18.5
opencv-python==4.2.0.34
openpyxl==3.0.3
pandas==1.0.4
parse==1.15.0
pbr==5.4.5
pdfminer.six==20200517
pycryptodome==3.9.7
pyee==7.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
pyppeteer==0.2.2
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-html==0.10.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.13.0
sqlparse==0.3.1
Tempita==0.5.2
tqdm==4.46.1
urllib3==1.25.9
w3lib==1.22.0
websockets==8.1
xlrd==0.7.1
xlwt==0.7.2
